pdf_printer/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
arrow==1.3.0
attrs==23.2.0
binaryornot==0.4.4
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cookiecutter==2.6.0
cycler==0.12.1
et-xmlfile==1.1.0
fastapi==0.114.2
flatbuffers==24.3.25
flet==0.21.2
flet-core==0.21.2
flet-runtime==0.21.2
fonttools==4.53.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.7
jax==0.4.28
jaxlib==0.4.28
Jinja2==3.1.4
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.4.0
MouseInfo==0.1.3
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.8.2
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytweening==1.2.0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
repath==0.9.0
requests==2.32.3
requests-futures==1.0.1
rich==13.8.1
scipy==1.13.1
sherlock-project==0.15.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.7
starlette==0.38.5
stem==1.8.2
text-unidecode==1.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
watchfiles==0.24.0
websockets==13.0.1